# Cache full script execution results in addition to signatures