@TheBlueMatt 2017-04-11
Fixed test_bitcoin segfaulting as it didnt init the script cache as it does the sigcache.