@TheBlueMatt 2017-04-12
Addressed Jeremy's comments aside from the request for a wrapper class, I think we need fewer dummy classes, not more :/.