@morcos 2017-04-25
Concept ACK
But I'd prefer if there were more safeguards in place against future changes that might cause consensus failure.  For instance, I think anything that is inputted to CScriptCheck should be committed to by the hash.  Right now it its the case that anything USED by CScriptCheck is committed to, but there is nothing stopping a future change to CScriptCheck that used the height from the Coins or something and caused problems.