@jtimon 2017-05-25
https://github.com/bitcoin/bitcoin/pull/10427 introduces GetScriptFlags like here but with some of my nits solved. If merged first, should make this a little bit smaller and simpler to review.
This needs rebase again.