@sdaftuar 2017-06-01
> This sounds like it would break (or at least complicate) CHECKBLOCKVERSION and possibly even CHECKBLOCKATHEIGHT...?

I'm skeptical that we'd ever adopt such context-dependent script flags (which strike me as a bad idea), but putting that aside: I benchmarked this patch on data from December 2016, and observed a 31% reduction in total block validation time, which is such a large performance improvement that I don't think this should be held up by a hypothetical type of consensus change, which we've not done before.  We can always revisit this performance optimization in the future if necessary.