@TheBlueMatt 2017-06-06
Addressed @sdaftuar's comments and rebased.