@TheBlueMatt 2017-06-09
Looks like it is to me?

On June 5, 2017 9:54:23 PM EDT, Pieter Wuille <notifications@github.com> wrote:
>This does not look like it's rebased on top of 10195?