@sipa 2017-06-09
@TheBlueMatt Now it is. When I made the comment, this PR had `CCoins` in it in some places.