@TheBlueMatt 2017-06-23
@sdaftuar pointed out that we could directly test CheckInputs' use of its own cache in unit tests, so I added a rather simple one that did so.