@TheBlueMatt 2017-06-27
I believe @sdaftuar indicated that he had a better test written for this, not sure what format it takes, however.