@TheBlueMatt 2017-06-27
Replaced test with one by @sdaftuar which is much better. Should be good to go now.